backend/pyproject.toml
[tool.poetry]
name = "django-admin-adapt"
version = "0.1.0"
description = ""
authors = ["ross <rossmere67@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.6"
psycopg2-binary = "^2.9.3"
strawberry-graphql = "^0.122.0"
strawberry-graphql-django = "^0.4.0"
strawberry-django-plus = "^1.21.0"
django-cors-headers = "^3.13.0"
django-extensions = "^3.2.0"
stringcase = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
